+++ src/commands/translation_commands/translate.react.js
@@ -10,6 +10,8 @@ const fn = require("../../core/helpers");
 const db = require("../../core/db");
 const logger = require("../../core/logger");
 const countryLangs = require("../../core/country.langs");
+const auth = require("../../core/auth");
+
 
 // ----------------------------------------------------
 // Translate a message through discord reaction (flag)
@@ -116,6 +118,7 @@ module.exports = function run (data, client)
                      delete data.message.attachments;
                      data.member.displayColor = fn.getRoleColor(data.message.member);
                      data.canWrite = true;
+                     data.reactuser = data.user_id;
 
                      // ------------------
                      // Start translation
@@ -133,6 +136,26 @@ module.exports = function run (data, client)
 
                      db.increaseStatsCount(col, id);
                      translate(data);
+                     if (server[0].flagpersist === false || server[0].flagpersist === 0)
+                     {
+
+                        try
+                        {
+
+                           setTimeout(() => data.message.reactions.resolve(emoji).users.remove(data.reactuser), auth.time.short);
+
+                        }
+                        catch (err)
+                        {
+
+                           console.log(
+                              "Command Message Deleted Error, translate.react.js = Line 152",
+                              err
+                           );
+
+                        }
+
+                     }
 
                   }
                );