+++ src/commands/utility_commands/help.js
@@ -6,11 +6,8 @@
 const sendMessage = require("../../core/command.send");
 const colors = require("../../core/colors");
 const discord = require("discord.js");
-const richEmbedMessage = new discord.MessageEmbed();
-const time = {
-   "long": 60000,
-   "short": 5000
-};
+const embed = new discord.MessageEmbed();
+const auth = require("../../core/auth");
 const helpFooter = `**For more help join our official [support server](<https://discord.gg/hXaedzCQ8d>)**\n\n`;
 
 // -------------
@@ -270,6 +267,9 @@ function helpMessage (config, param)
    "```md\n" +
    `# Command\n` +
    `* ${cmd} react [on/off] \n\n` +
+   `# Reaction Translation auto delete\n` +
+   `* ${cmd} settings reactpersist [on/off]\n` +
+   `* ${cmd} settings flagpersist [on/off]\n` +
    "```" +
 
    "```md\n" +
@@ -456,27 +456,52 @@ function helpMessage (config, param)
    `__**Settings**__\n\n` +
 
    "```md\n" +
-   `# Set default server language\n` +
-   `* ${cmd} settings setLang to [lang]\n\n` +
+   `# Current Server Settings\n` +
+   `* ${cmd} settings\n\n` +
    "```" +
 
    "```md\n" +
-   `# Style settings\n` +
-   `* ${cmd} embed [on/off]\n\n` +
-
-   `# Parameters\n` +
-   `* on - Turns on Embed Translation\n` +
-   `* off - Turns on Webhook Translation Sending\n\n` +
-
-   `# Examples\n` +
-   `* ${cmd} embed on \n` +
-   `* ${cmd} embed off \n` +
+   `# Set Custom Bot Prefix\n` +
+   `* ${cmd} prefix [prefix]\n` +
+   `* ${long} prefix [prefix]\n\n` +
    "```" +
 
    "```md\n" +
-   `# Help menu auto deletion\n` +
-   `* ${cmd} settings persist [on/off]\n` +
-   "```\n";
+   `# Developer Announcment Messages\n` +
+   `* ${cmd} announce [on/off]\n` +
+   "```" +
+
+   "```md\n" +
+   `# Embedded Message Style\n` +
+   `* ${cmd} embed [on/off]\n` +
+   `* ${cmd} help embed\n` +
+   "```" +
+
+   "```md\n" +
+   `# Bot to Bot Translation Status\n` +
+   `* Usually 90% of bots ignore other bot messages but this feature attemptes to translate them.\n` +
+   `* ${cmd} bot2bot [on/off]\n` +
+   `* ${cmd} help bot2bot\n` +
+   "```" +
+
+   "```md\n" +
+   `# Reaction Translations\n` +
+   `* ${cmd} react [on/off]\n` +
+   "```" +
+
+   "```md\n" +
+   `# Help Menu auto delete\n` +
+   `* ${cmd} settings persist [on/off]\n\n` +
+   `# Reaction Translation auto delete\n` +
+   `* ${cmd} settings reactpersist [on/off]\n\n` +
+   `# Flag Emojoi auto delete\n` +
+   `* ${cmd} settings flagpersist [on/off]\n\n` +
+   "```" +
+
+   "```md\n" +
+   `# Webhook Debug Active State\n` +
+   `* ${cmd} debug [on/off]\n\n` +
+   "```";
 
    // -------------------
    // Statistics Command
@@ -754,11 +779,22 @@ module.exports = function run (data)
       {
 
          console.log("Insufficient Permission");
-         data.message.delete({"timeout": time.short}).catch((err) => console.log(
-            "Command Message Deleted Error, help.js = ",
-            err
-         ));
-         richEmbedMessage.
+         try
+         {
+
+            setTimeout(() => data.message.delete(), auth.time.short);
+
+         }
+         catch (err)
+         {
+
+            console.log(
+               "Command Message Deleted Error, help.js = Line 770",
+               err
+            );
+
+         }
+         embed.
             setColor(colors.get(data.color)).
             setDescription("This command is available only to Developers. \n\n").
             setTimestamp().
@@ -768,13 +804,24 @@ module.exports = function run (data)
          // Send message
          // -------------
 
-         return data.message.channel.send(richEmbedMessage).then((msg) =>
+         return data.message.channel.send(embed).then((msg) =>
          {
 
-            msg.delete({"timeout": time.long}).catch((err) => console.log(
-               "Bot Message Deleted Error, help.js = ",
-               err
-            ));
+            try
+            {
+
+               setTimeout(() => msg.delete(), auth.time.short);
+
+            }
+            catch (err)
+            {
+
+               console.log(
+                  "Command Message Deleted Error, help.js = Line 798",
+                  err
+               );
+
+            }
 
          });