+++ src/core/db.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-undef */
 // -----------------
 // Global variables
 // -----------------
@@ -7,6 +6,7 @@
 /* eslint-disable sort-keys */
 /* eslint-disable no-unused-vars */
 /* eslint-disable quote-props */
+/* eslint-disable no-undef */
 const autoTranslate = require("./auto");
 const Sequelize = require("sequelize");
 const logger = require("./logger");
@@ -180,6 +180,14 @@ const Servers = db.define(
          "defaultValue": false
       },
       "flag": {
+         "type": Sequelize.BOOLEAN,
+         "defaultValue": true
+      },
+      "flagpersist": {
+         "type": Sequelize.BOOLEAN,
+         "defaultValue": true
+      },
+      "reactpersist": {
          "type": Sequelize.BOOLEAN,
          "defaultValue": true
       },
@@ -528,6 +536,8 @@ exports.updateColumns = async function updateColumns ()
    await this.addTableColumn("servers", serversDefinition, "warncount", Sequelize.INTEGER, 0);
    await this.addTableColumn("servers", serversDefinition, "ejectcount", Sequelize.INTEGER, 0);
    await this.addTableColumn("servers", serversDefinition, "flag", Sequelize.BOOLEAN, true);
+   await this.addTableColumn("servers", serversDefinition, "reactpersist", Sequelize.BOOLEAN, true);
+   await this.addTableColumn("servers", serversDefinition, "flagpersist", Sequelize.BOOLEAN, true);
    // console.log("DEBUG: All Columns Checked or Added");
 
    // For older version of RITA, must remove old unique index
@@ -709,8 +719,7 @@ exports.removeTask = function removeTask (origin, dest, cb)
 
       // console.log("DEBUG: removeTask() - all");
       return Tasks.destroy({"where": {[Op.or]: [
-         {origin},
-         {"dest": origin}
+         {origin}
       ]}}).then(function error (err, result)
       {
 
@@ -724,9 +733,7 @@ exports.removeTask = function removeTask (origin, dest, cb)
    }
    return Tasks.destroy({"where": {[Op.or]: [
       {origin,
-         dest},
-      {"origin": dest,
-         "dest": origin}
+         dest}
    ]}}).then(function error (err, result)
    {