+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 * Local hosted users persist and react command fixed.
 * CS same language now works.
 * Fixed invalid channel message.
+* `!tr reactpersist` has been added. - Reaction will auto delete after 60 seconds.
+* `!tr flagpersist` has been added. - Flag emoji will be removed after 15 seconds.
+* Help menu updated.
 
 ## 1.2.6 Change Log
 * RITA is now at 6400 servers, This is just Amazing.